I think we're finall done

+++ docker/database/init.sql
@@ -36,7 +36,7 @@ CREATE TABLE StudentExams (
     studentexam_id INT AUTO_INCREMENT NOT NULL,
     user_id INT NOT NULL,
     exam_id INT NOT NULL,
-    score INT NOT NULL DEFAULT -1,
+    score FLOAT NOT NULL DEFAULT -1,
     date datetime NOT NULL,
     FOREIGN KEY(user_id) REFERENCES Users(user_id),
     FOREIGN KEY(exam_id) REFERENCES Exams(exam_id),
@@ -46,7 +46,7 @@ CREATE TABLE StudentExams (
 CREATE TABLE ExamQuestions (
     studentexam_id INT NOT NULL,
     question_id INT NOT NULL,
-    points INT NOT NULL,
+    points FLOAT NOT NULL,
     answer VARCHAR(2000),
     comment VARCHAR(500),
     FOREIGN KEY(studentexam_id) REFERENCES StudentExams(studentexam_id),
@@ -67,8 +67,8 @@ CREATE TABLE Autograde (
     studentexam_id INT NOT NULL,
     testcase_id INT NOT NULL,
     autoresult VARCHAR(300) NOT NULL,
-    points INT NOT NULL,
-    score INT NOT NULL DEFAULT 0,
+    points FLOAT NOT NULL,
+    score FLOAT NOT NULL DEFAULT 0,
     FOREIGN KEY(studentexam_id) REFERENCES StudentExams(studentexam_id),
     FOREIGN KEY(testcase_id) REFERENCES Testcases(testcase_id),
     PRIMARY KEY(studentexam_id, testcase_id)
